+++ eas-20210701/alibabacloud_eas20210701/models/_list_resource_instances_request.py
@@ -20,6 +20,7 @@ class ListResourceInstancesRequest(DaraModel):
         page_number: int = None,
         page_size: int = None,
         sort: str = None,
+        zone: str = None,
     ):
         # The billing method of the instance. Valid values:
         # 
@@ -220,6 +221,7 @@ class ListResourceInstancesRequest(DaraModel):
         # 
         #     <!-- -->
         self.sort = sort
+        self.zone = zone
 
     def validate(self):
         pass
@@ -262,6 +264,9 @@ class ListResourceInstancesRequest(DaraModel):
         if self.sort is not None:
             result['Sort'] = self.sort
 
+        if self.zone is not None:
+            result['Zone'] = self.zone
+
         return result
 
     def from_map(self, m: dict = None):
@@ -298,6 +303,9 @@ class ListResourceInstancesRequest(DaraModel):
 
         if m.get('Sort') is not None:
             self.sort = m.get('Sort')
+
+        if m.get('Zone') is not None:
+            self.zone = m.get('Zone')
 
         return self